+++ FluentSecurity/PolicyContainerExtensions.cs
@@ -31,16 +31,16 @@ namespace FluentSecurity
 			return policyContainer;
 		}
 
-		public static IPolicyContainerConfiguration DenyAnonymousAccess(this IPolicyContainerConfiguration policyContainer)
+		public static IPolicyContainerConfiguration<DenyAnonymousAccessPolicy> DenyAnonymousAccess(this IPolicyContainerConfiguration policyContainer)
 		{
 			policyContainer.AddPolicy(new DenyAnonymousAccessPolicy());
-			return policyContainer;
+			return new PolicyContainerConfigurationWrapper<DenyAnonymousAccessPolicy>(policyContainer);
 		}
 
-		public static IPolicyContainerConfiguration DenyAuthenticatedAccess(this IPolicyContainerConfiguration policyContainer)
+		public static IPolicyContainerConfiguration<DenyAuthenticatedAccessPolicy> DenyAuthenticatedAccess(this IPolicyContainerConfiguration policyContainer)
 		{
 			policyContainer.AddPolicy(new DenyAuthenticatedAccessPolicy());
-			return policyContainer;
+			return new PolicyContainerConfigurationWrapper<DenyAuthenticatedAccessPolicy>(policyContainer);
 		}
 
 		public static IPolicyContainerConfiguration Ignore(this IPolicyContainerConfiguration policyContainer)
@@ -49,16 +49,16 @@ namespace FluentSecurity
 			return policyContainer;
 		}
 
-		public static IPolicyContainerConfiguration RequireRole(this IPolicyContainerConfiguration policyContainer, params object[] roles)
+		public static IPolicyContainerConfiguration<RequireRolePolicy> RequireRole(this IPolicyContainerConfiguration policyContainer, params object[] roles)
 		{
 			policyContainer.AddPolicy(new RequireRolePolicy(roles));
-			return policyContainer;
+			return new PolicyContainerConfigurationWrapper<RequireRolePolicy>(policyContainer);
 		}
 
-		public static IPolicyContainerConfiguration RequireAllRoles(this IPolicyContainerConfiguration policyContainer, params object[] roles)
+		public static IPolicyContainerConfiguration<RequireAllRolesPolicy> RequireAllRoles(this IPolicyContainerConfiguration policyContainer, params object[] roles)
 		{
 			policyContainer.AddPolicy(new RequireAllRolesPolicy(roles));
-			return policyContainer;
+			return new PolicyContainerConfigurationWrapper<RequireAllRolesPolicy>(policyContainer);
 		}
 	}
 }